Xóa các comment nhỏ bên trong nội dung các hàm và cải thiện hệ thống phân quyền

+++ app/Http/Controllers/Api/UserController.php
@@ -10,6 +10,8 @@ use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Illuminate\Auth\Access\AuthorizationException;
+use Bouncer;
 
 class UserController extends Controller
 {
@@ -33,9 +35,14 @@ class UserController extends Controller
      *
      * @param Request $request
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function index(Request $request): JsonResponse
     {
+        if (!Bouncer::can('view-users')) {
+            throw new AuthorizationException('Bạn không có quyền xem danh sách người dùng.');
+        }
+
         $perPage = $request->query('per_page', 15);
         $filters = $request->only(['search']);
         
@@ -49,9 +56,14 @@ class UserController extends Controller
      *
      * @param int $id
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function show(int $id): JsonResponse
     {
+        if (!Bouncer::can('view-users')) {
+            throw new AuthorizationException('Bạn không có quyền xem thông tin người dùng.');
+        }
+
         try {
             $user = $this->userService->getUserById($id);
             
@@ -66,9 +78,14 @@ class UserController extends Controller
      *
      * @param CreateUserRequest $request
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function store(CreateUserRequest $request): JsonResponse
     {
+        if (!Bouncer::can('create-users')) {
+            throw new AuthorizationException('Bạn không có quyền tạo người dùng mới.');
+        }
+        
         $userData = $request->validated();
         
         $user = $this->userService->createUser($userData);
@@ -82,9 +99,14 @@ class UserController extends Controller
      * @param UpdateUserRequest $request
      * @param int $id
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function update(UpdateUserRequest $request, int $id): JsonResponse
     {
+        if (!Bouncer::can('update-users')) {
+            throw new AuthorizationException('Bạn không có quyền cập nhật thông tin người dùng.');
+        }
+        
         try {
             $userData = $request->validated();
             
@@ -101,9 +123,14 @@ class UserController extends Controller
      *
      * @param int $id
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function destroy(int $id): JsonResponse
     {
+        if (!Bouncer::can('delete-users')) {
+            throw new AuthorizationException('Bạn không có quyền xóa người dùng.');
+        }
+        
         try {
             $deleted = $this->userService->deleteUser($id);
